+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Version - 0]
+## Feature [v0.28.0] 2023-05-19
+- added planning point menu
 ## Fix [v0.27.4] 2023-05-17
 - fixed logic delete & edit in campaign point menu
 ## Fix [v0.27.3] 2023-05-17